# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Spec Doc API
  version: "1.0.0"
  description: >
    Phase 1 contract — health check, projects CRUD, and context
    read/write for four named keys (builder, principles, codebase,
    references). This file is the build artifact; flask/dtos/ and
    mock_server.py are derived from it.
servers:
  - url: http://localhost:3101
    description: Flask development server
  - url: http://localhost:3102
    description: Mock server (Task 3)
paths:
...
  /api/projects:
    get:
      summary: List all projects, newest first
      operationId: listProjects
      responses:
        "200":
          description: Array of project summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/projects/{project_id}/generate-task:
    post:
      summary: Spawn a background thread to generate the next missing implementation guide
      operationId: startGenerateTask
      tags: [task_gen]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Background generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateTaskStartResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A generation thread is already running for this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateTaskStartResponse"
        "429":
          $ref: "#/components/responses/UsageLimitReached"
  /api/projects/{project_id}/generate-task/status:
    get:
      summary: Read the in-process state of the background generate-task thread
      operationId: getGenerateTaskStatus
      tags: [task_gen]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current generation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateTaskStatusResponse"
        "502":
          description: Linter blocked the write — one or more error-severity flags present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateTaskStatusResponse"
...
  /api/projects/{project_id}/cancel:
    post:
      summary: Request cooperative cancellation of an in-flight task_gen job
      operationId: cancelTaskGen
      tags: [task_gen]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Cancellation requested; status flips to CANCELLING.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelAcceptedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job is not in a cancellable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/projects/{project_id}/regenerate-task:
    post:
      summary: Delete and re-run a previously generated task guide
      operationId: regenerateTask
      tags: [task_gen]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateTaskRequest'
      responses:
        '202':
          description: Regeneration started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTaskStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A generation thread is already running for this slot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTaskStartResponse'
        '429':
          $ref: '#/components/responses/UsageLimitReached'
...
components:
  schemas:
...
    SpecSummary:
      type: object
      required: [filename, label]
      properties:
        filename:
          type: string
        label:
          type: string
...
    ProjectSummary:
      type: object
      required: [id, name, createdAt, specs]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        specs:
          type: array
          items:
            $ref: "#/components/schemas/SpecSummary"
...
    ProjectCreateRequest:
      type: object
      required: [name, files]
      properties:
        name:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileInput"
    FileInput:
      type: object
      required: [filename, content]
      properties:
        filename:
          type: string
        content:
          type: string
    ProjectCreateResponse:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CancelAcceptedResponse:
      type: object
      description: >
        Returned with HTTP 202 by the cancel routes once the
        WorkflowExecution has been flipped to CANCELLING. Final
        CANCELLED transition happens between step boundaries on the
        runtime's next iteration.
      required: [status]
      properties:
        status:
          type: string
          description: Current execution status string after the request.
          example: CANCELLING
...
    RegenerateTaskRequest:
      type: object
      required: [task_num]
      properties:
        task_num:
          type: string
          description: >
            Task number whose generated guide should be deleted and
            re-generated. Matches the num column in the epic backlog.
...
    GenerateTaskStartResponse:
      type: object
      required: [started]
      properties:
        started:
          type: boolean
          description: True if a new background thread was spawned, false if one was already running.
        alreadyRunning:
          type: boolean
          description: Present and true when a generation thread for this project was already running.
    GenerateTaskStatusResponse:
      type: object
      required: [running, done]
      properties:
        running:
          type: boolean
          description: True while the background thread is in progress.
        done:
          type: boolean
          description: True once the thread has terminated (successfully or not).
        allDone:
          type: boolean
          description: Present and true when every task in the epic already has a matching guide on disk.
        filename:
          type: string
          description: Name of the task-N-*.md file that was written (only when done and not allDone).
        taskNum:
          type: string
          description: Task number from the epic table (e.g., "2").
        taskName:
          type: string
          description: Task name from the epic table.
        error:
          type: string
          description: Error message captured from the background thread.
        lintErrors:
          type: array
          description: >
            Present when error-severity lint flags blocked the file write.
            The polling route returns 502 when this field is present.
          items:
            $ref: "#/components/schemas/LintFlag"
        warnings:
          type: array
          description: >
            Present when warning-severity lint flags were raised but the
            file was still written successfully.
          items:
            $ref: "#/components/schemas/LintFlag"
    LintFlag:
      type: object
      required: [rule, severity, message]
      properties:
        rule:
          type: string
          description: Lint rule identifier (e.g. "personal-path", "preamble-leak").
        severity:
          type: string
          enum: [error, warning]
          description: '"error" blocks the write; "warning" allows it with the flag surfaced.'
        message:
          type: string
          description: Human-readable violation description.
        line:
          type: integer
          nullable: true
          description: 1-based line number of the first violation site, if known.
...
    UsageLimitError:
      type: object
      description: >
        Returned with HTTP 429 by any AI route gated by check_usage_limit
        when a free-tier user has exhausted their daily quota for that
        feature. The Angular usage-limit interceptor catches this body
        globally and routes the user to upgrade_url.
      required:
        - error
        - feature
        - limit
        - reset_at
        - upgrade_url
      properties:
        error:
          type: string
          description: Stable error code for client dispatch.
          example: free_tier_limit_reached
        feature:
          type: string
          enum: [bootstrap, task_gen, spec_gen]
          description: >
            Feature key that hit the daily cap. Matches keys in
            DAILY_FREE_TIER_LIMITS in modules/usage/service.py.
          example: task_gen
        limit:
          type: integer
          minimum: 0
          description: >
            Daily cap for this feature on the free tier. Free-tier caps are
            bootstrap=3 / task_gen=20 / spec_gen=10 per UTC day.
          example: 20
        reset_at:
          type: string
          format: date-time
          description: UTC midnight at which the daily counter resets.
          example: "2026-04-27T00:00:00Z"
        upgrade_url:
          type: string
          description: >
            App-relative path to the upgrade page. Angular's 429 interceptor
            navigates to this URL when a UsageLimitError is received.
          example: /upgrade
...
  responses:
    BadRequest:
      description: Missing or invalid request fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UsageLimitReached:
      description: >
        Daily free-tier limit reached for the gated feature. Free-tier caps are
        bootstrap=3 / task_gen=20 / spec_gen=10 per UTC day. Pro users never
        receive this response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UsageLimitError"
          example:
            error: free_tier_limit_reached
            feature: task_gen
            limit: 20
            reset_at: "2026-04-27T00:00:00Z"
            upgrade_url: /upgrade
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Neon Auth JWT issued at login. Pass as Authorization: Bearer <token>.
        Verified by the require_auth decorator (implemented in Task 2).